refactor: move ClientCredentialsSession to zenpy/oauth.py, drop Zenpy-side client_id/secret/scope handling
Addresses facetoe's review feedback:
- Move ClientCredentialsSession out of __init__.py into zenpy/oauth.py
- Remove client_id/client_secret/scope/expires_in from Zenpy.__init__,
  along with the auto-construction and partial-args validation logic.
  Users now build a ClientCredentialsSession themselves and pass it
  via the session param.
- Update usage examples in README.md and docs/zenpy.rst accordingly
- Update imports and tests in tests/test_api/test_client_credentials.py

zenpy/__init__.py is now identical to its state before this PR.
